Create New Habit - Kotlin

+++ Code/app/src/main/java/com/example/habittracker/Habit.kt
@@ -1,4 +1,4 @@
 package com.example.habittracker
 
 data class Habit(val id : Int, var name : String, var description : String,
-                 var priority : Int, var remindTime : String)
+                 var priority : Float, var remindTime : String)
